+++ source/packages/@aws-lza/interfaces/control-tower/setup-landing-zone.ts
@@ -137,6 +137,12 @@ export interface ISetupLandingZoneConfiguration {
      */
     readonly audit: ISharedAccountDetails;
   };
+  /**
+   * Flag indicating whether account auto-enrollment is enabled or not.
+   * When enabled, accounts moved between OUs are automatically enrolled in Control Tower.
+   * Available for Landing Zone version 3.1 and above.
+   */
+  readonly accountAutoEnrollment?: boolean;
 }
 
 /**