+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat(pipeline): replace codestar notification with EventBridge and SNS
 - feat(security-hub): added support for AWS Security Hub automation rules
 - feat(control-tower): stop enrolling accounts with Service Catalog
+- feat(control-tower): support auto account enrollment
 
 ### Fixed